requirements.txt
black==23.3.0
boto3==1.36.11
botocore==1.36.11
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Flask==2.2.5
idna==3.7
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
mongoengine==0.29.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.8.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.2.3
zipp==3.21.0
roboflow==1.1.53